requirements.txt
numpy==1.23.3
Pillow==9.3.0
streamlit==1.15.1
tensorflow==2.11.0
